:recycle: refactor: 타입 수정

+++ hugme-dog/src/main/java/spring/hugme/domain/community/dto/response/CommentWriteResponse.java
@@ -22,6 +22,6 @@ public class CommentWriteResponse {
 
   LocalDateTime updatedAt;
 
-  int commentCount;
+  Long commentCount;
 
 }
